@@ -14,7 +14,9 @@ extension PHImageManager {
         targetSize: CGSize,
         contentMode: PHImageContentMode,
         options: PHImageRequestOptions?
-    ) async -> (UIImage?, [AnyHashable : Any]?) {
+    ) async -> UIImage? {
+        options?.isSynchronous = false
+        
         var callCount = 0
         
         return await withCheckedContinuation { continuation in
@@ -23,10 +25,69 @@ extension PHImageManager {
                 targetSize: targetSize,
                 contentMode: contentMode,
                 options: options
-            ) { image, dict in
+            ) { image, _ in
                 guard callCount == 0 else { return }
                 callCount += 1
-                continuation.resume(returning: (image, dict))
+                continuation.resume(returning: image)
+            }
+        }
+    }
+}
+
+
+// TODO: - Start testing and using
+extension PHImageManager {
+    
+    public enum ImageRequestError: Error {
+        case cancelled
+        case imageDataUnavailable
+        case underlyingError(Error)
+    }
+
+    public func requestImageV2(
+        for asset: PHAsset,
+        targetSize: CGSize,
+        contentMode: PHImageContentMode,
+        options: PHImageRequestOptions? = nil
+    ) async throws -> UIImage {
+        
+        return try await withCheckedThrowingContinuation { continuation in
+            requestImage(
+                for: asset,
+                targetSize: targetSize,
+                contentMode: contentMode,
+                options: options
+            ) { image, info in
+                guard let info else {
+                    continuation.resume(throwing: ImageRequestError.imageDataUnavailable)
+                    return
+                }
+                
+                // Check if the request was canceled.
+                if let isCancelled = info[PHImageCancelledKey] as? Bool, isCancelled {
+                    continuation.resume(throwing: ImageRequestError.cancelled)
+                    return
+                }
+                
+                // Check for errors in the info dictionary.
+                if let error = info[PHImageErrorKey] as? Error {
+                    continuation.resume(throwing: ImageRequestError.underlyingError(error))
+                    return
+                }
+                
+                // Check if the image is degraded. If it is, ignore it.
+                if let isDegraded = info[PHImageResultIsDegradedKey] as? Bool, isDegraded {
+                    return // Ignore degraded image, wait for full-quality image.
+                }
+                
+                // Ensure the full-quality image is available.
+                guard let image = image else {
+                    continuation.resume(throwing: ImageRequestError.imageDataUnavailable)
+                    return
+                }
+                
+                // Resume with the full-quality image.
+                continuation.resume(returning: image)
             }
         }
     }
